Added an event bus for application wide communication.
You can use the bus to emit messages with
bus.$emit('messageName', messageValue)
and listen where you like with
bus.$on('messageName, callBack(messageValue) {...})
Application-wide data and methods can also be stored there. I've added the connection, and a setter for it. Generally, bus data shouldn't be set directly since that gets back to "whose doing what to it?"
This will probably be replaced when we get to Vuex.